[1] naming and code organization improvement
