+++ GNUmakefile
@@ -1,5 +1,5 @@
 ################################################################################
-#  Copyright (C) 2016  Sergey Morozov <sergey94morozov@gmail.com>              #
+#  Copyright (C) 2016, 2017  Sergey Morozov <sergey94morozov@gmail.com>        #
 #                                                                              #
 #  This program is free software: you can redistribute it and/or modify        #
 #  it under the terms of the GNU General Public License as published by        #
