--- src/lib/ops.h
@@ -1,32 +1,0 @@
-/**
- * Copyright (C) 2016  Sergey Morozov <sergey94morozov@gmail.com>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#ifndef OPS_H
-#define OPS_H
-
-#include <cloudtiering.h>
-
-/**
- * S3 PROTOCOL
- */
-
-int  s3_init_remote_store_access(void);
-int  s3_move_file_in(const char *path);
-int  s3_move_file_out(const char *path);
-void s3_term_remote_store_access(void);
-
-#endif /* OPS_H */
